Add test for invalid job ID
