+++ tests/base.py
@@ -1,5 +1,7 @@
 import os
+import sys
 import unittest
+import time
 
 from jasmin_arc import ArcInterface
 
@@ -8,7 +10,6 @@ class ArcTestCase(unittest.TestCase):
     """
     Base class for tests - used to create an ArcInterface instance on class set up
     """
-
     @classmethod
     def setUpClass(cls):
         config_path = os.environ.get("JASMIN_ARC_CONFIG", None)
@@ -18,3 +19,14 @@ class ArcTestCase(unittest.TestCase):
             print("Using default config")
 
         cls.ARC_INTERFACE = ArcInterface(config_path)
+
+    def wait(self, n):
+        """
+        Wait for n seconds
+        """
+        sys.stdout.write("Waiting for {} seconds".format(n))
+        for _ in range(n):
+            sys.stdout.write(".")
+            sys.stdout.flush()
+            time.sleep(1)
+        sys.stdout.write("\n")
